Use app_data_file_type for mls rules.
Now that we have an attribute for all app data files, make use of
it. It's cleaner.

The net effect here is a slight loosening of permissions - we now
allow open fds for any app_data_file_type to be passed to a different
process, rather than just app_data_file and privapp_data_file.

Bug: 171795911
Test: presubmits
Change-Id: I4cf812d01577b923efbe1ea3f276c209844d8858