[DO NOT SUBMIT] Add GWP-ASan sysprops sepolicy.
These flags should be writeable to the shell for both root and non-root
users. They should be readable everywhere, as they're read in libc
during initialization (and there's nothing secret to hide). We just
don't want to allow apps to set these properties.

These properties are non-persistent, are for local developer debugging
only.

Bug: N/A
Test: N/A
Change-Id: I478e364efe24d2629a3db03dd978ab0f30ef4a47