Remove fuse.passthrough.enabled get_prop scope from untrusted_app
To fix a presubmit issue with GSI mixed builds, this new boolean sysprop
fuse.passthrough.enabled has been introduced, and changes the test
behavior depending on that.

As long as the test is run as untrusted_app, getting the value of that
property requires the sepolicy rule to be set as readable for the whole
appdomain.
This was a temporary solution and now that the test has been refactored,
untrusted_app can be removed from the get_prop scope of
fuse.passthrough.enabled.

Bug: 187912701
Bug: 186635810
Test: CtsScopedStorageDeviceOnlyTest
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: Ib72a8edf8299ba5c70f25f4192f893b1d370dde3