cts case fail about lack permission
The implementation of gralloc of s9863a1h10 GPU is special. It uses DRM to register its device nodes. It creates nodes card1 and renderD128, so that other apps/services need to have access to these nodes when accessing gralloc.

We need to allow the mediatranscoding in the system partition to have sufficient permissions to access the 3 nodes (card0, card1 and renderD128) to avoid the following errors.

Test: run the cts case in s9863a1h10 of 2g, and see if the case will fail.

Bug: 199581284
Change-Id: I88393ea60b98f7b4636907e8e163c011ae11b5ea