Introduce ro.boot.hypervisor properties
After discussions, a Hypervisor HAL is too heavy an entity for the current
goal, i.e. representing a host version string and whether Android itself
is acting as a VM orchestrator vs. a guest.

Represent this by a pair of new system properties:
 * ro.boot.hypervisor.android_host
   a boolean property representing whether Android is the host system;
 * ro.boot.hypervisor.version
   a free-form descriptor of the current host/hypervisor version

These properties are meant to be provided to Android as androidboot.
properties by whatever host-specific means are used to supply other boot
properties to the target Android instance. Access could be later opened
to other vendor processes to set if needed for specific setups where init
is not a sufficiently-early stage for host/guest communication. Such setups
are not known at this time.

For a native Android incantation, they default to being missing, which should
be interpreted as "native, not acting as VM host".

Bug: 178749018
Test: build cuttlefish
Change-Id: Id721c14ef1958b525c2866a660dcae8fd176a79d