allow system service to fork exec clatd
we are planning to move clat into mainline module. in order to do
that, we need:
(a) system server being able to fork exec clatd
(b) clatd binary being setuid/setgid

for a), we need this security rule; for b), it requires apex support
which is working on b/209971551.

Test: manual test which launches clatd from system server and check
that ipv4 traffic can work under ipv6 only network.
- ping 8.8.8.8
- testipv4.com shows 10/10

Change-Id: Ic31e7005ebcf7c0ae2865551dac6ce41c61ca4dd