Allow composd to create odrefresh staging directory
composd in responsible to prepare the staging directory for odrefresh
(in the VM) to write the output to. The directory must be set up
correctly, through the common palette function,
PaletteCreateOdrefreshStagingDirectory (thus these privileges), to be
compatible with early boot compilation. This should make it the easiest
for odsign to deal with.

Bug: 205750213
Test: No denials

Change-Id: I9444470b31518242c1bb84fc755819d459d21d68