Adding ro.adb.allow_root
To automate USER build testing, some partners are overriding the
following properties to disable ADB authentication:

  - ro.adb.secure=0
  - ro.debuggable=1

However, when running `adb root`, the adb connection gets
stuck because it lacks the userdebug sepolicy. Adding a new
property "ro.adb.allow_root" and set it to 0 to disallow
`adb root` to avoid this issue for this scenario.

Bug: 208163696
Test: Runs `adb root` with "ro.adb.allow_root" is 0
      => "adbd cannot run as root when ro.adb.allow_root is false"
Change-Id: I0ddd6928d9f8daeba18e67f1e7701055aa930438