Add deprecated_soc_prop
Some system properties have previously been used by vendors to expose
information about the SoC. With the new limitations in place on the
namespace available to vendor properties, it is not possible to define
them in the vendor scope anymore. Define them as part of the platform so
it is still possible to add a rule on them (such as dontaudit).

Bug: 210056469
Test: lunch aosp_bramble-userdebug && m
Change-Id: I7fe53b875618a704ef3ada6312faa6e8b390baa6