Adds GPU sepolicy for devices with DRM-based gralloc/rendering
... such as Cuttlefish (Cloud Android virtual device) which has a
DRM virtio-gpu based gralloc and (sometimes) DRM virtio-gpu based
rendering (when forwarding rendering to the host machine with
mesa in the guest and virgl_renderer in the host).

See the connected change aosp/1997572 for what was previously
needed for Cuttlefish and which seems to be replicated for
device/linaro/dragonboard as well.

The `sysfs_gpu:file` comes from Mesa using libdrm's
`drmGetDevices2()` which calls into `drmParsePciDeviceInfo()` to
get vendor id, device id, version etc.

Bug: b/161819018
Test: launch_cvd
Test: launch_cvd --gpu_mode=gfxstream
Change-Id: Ie6ebd02087e1fc949dc0f34f7f93cda6ee6e6690