surfaceflinger: allow to be profiled on release builds
This is a key system process for certain performance investigations, so
allow profiling of its native heap and general callstack sampling. This
is already allowed on debuggable builds via domain.te.

In addition to the sepolicy, the profiler itself does checks on whether
to allow profiling. At the time of writing, profiling requests coming
from "shell" for surfaceflinger will be disallowed (as it is a native
process running as "system"). However profiling requests coming from the
platform via "statsd" will be allowed.

Change-Id: Ib092064ea911aed08d981adc823cd871fc271a96