Revert "Add sepolicy for bluetooth.core.gap.le.conn.min.limit sy..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/e2a95915-5566-48c2-a0f4-436322ad6a98). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idafb7ff6043ef781b53b9a68a6ba1ba83ad7e703